Updated Assignment2Application
Based on suggestions from the code review. I removed the comments. Removed some line breaks. Created a constant for the max amount of guesses. Moved the "Pick a number..." statement to cut down on redundancy. Changed the arrangement of the logic in the while loop. Removed the final else statement and closed the scanner.